[dgarciabravo] Buscador Express

+++ siga-core/src/main/java/org/itcgae/siga/db/services/cen/mappers/CenPersonaExtendsMapper.java
@@ -175,7 +175,7 @@ public interface CenPersonaExtendsMapper extends CenPersonaMapper{
 	@SelectProvider(type = CenPersonaSqlExtendsProvider.class, method = "busquedaColegiadoExpress")//falta hacer query
 	@Results({
 		@Result(column = "NCOLEGIADO", property = "nColegiado", jdbcType = JdbcType.VARCHAR),
-		@Result(column = "IDPERSONA", property = "idPersona", jdbcType = JdbcType.NUMERIC),
+		@Result(column = "IDPERSONA", property = "idPersona", jdbcType = JdbcType.BIGINT),
 		@Result(column = "NOMBRE", property = "nombre", jdbcType = JdbcType.VARCHAR),
 	})
 	List<ColegiadoJGItem> busquedaColegiadoExpress(String colegiadoJGItem, String idInstitucion);
